fix: Correct WSGI path for Gunicorn in Dockerfile
The WSGI application path in the Gunicorn command within the Dockerfile
was incorrect. It pointed to `config.wsgi:application`, which is
incorrect.  This commit updates it to `core.wsgi:application` to
correctly point to the WSGI application entrypoint within the Django
project's `core` directory. This resolves a deployment issue where the
application would fail to start due to Gunicorn not finding the WSGI
application.