feat: Add initial LoanApplication model definition
This commit introduces the `LoanApplication` model class to
represent loan application data. This is the initial setup
and does not yet include any specific fields. The model will
be further populated with relevant properties in subsequent
commits. This model will be used for handling loan application
data throughout the loan application process.