feat: Add file_selector_linux plugin to generated plugins
This commit adds the `file_selector_linux` plugin to the
`FLUTTER_PLUGIN_LIST` in `generated_plugins.cmake`. This ensures
that the `file_selector_linux` plugin is correctly included and
linked in the Linux build of the Flutter application. The plugin
provides file selection functionality, enabling the app to interact
with the operating system's file system for tasks such as opening
and saving files.