feat: Register FileSelectorWindows plugin
This commit adds the FileSelectorWindows plugin registration
to the `generated_plugin_registrant.cc` file. This enables
the use of file selection functionality within the Windows
version of the application.

The `FileSelectorWindowsRegisterWithRegistrar` function is
called to register the plugin with the Flutter plugin registry.
This ensures that the plugin is properly initialized and can
be accessed by the Flutter application.