feat: Add jwt-decode package for JWT token handling
This commit adds the `jwt-decode` package as a dependency. This package will be used for decoding JSON Web Tokens (JWTs) on the frontend, allowing us to extract user information and other relevant data stored within the token. This is necessary for implementing authentication and authorization features. The specific version added is 4.0.0.