feat: Add React logo SVG asset for visual enhancements
This commit introduces 'react.svg', the React logo, to the assets directory.
This SVG file will be used to visually represent React in the frontend,
enhancing the user interface and providing better context within the application.
The logo is added to improve the overall look and feel of the sacco frontend.

Keywords: React, logo, SVG, asset, frontend, UI, visual