feat: Implement initial LoanList component structure (empty)
This commit introduces the LoanList component, laying the foundation
for displaying a list of loans. Currently, the component is a placeholder
and doesn't contain any actual loan listing functionality. Subsequent
commits will populate this component with data and interactive elements.
This component will be used to show the loan data to the users.