feat: Add PublicRoute component for unauthenticated access
This commit introduces the `PublicRoute` component. Currently, it's
a placeholder as indicated by the empty file. This component will be
used to wrap routes that should only be accessible to users who are
not authenticated, redirecting authenticated users away from these
pages. This sets the foundation for implementing authentication-based
route protection in the application.